correction de permutation

+++ coordinates.h
@@ -12,8 +12,8 @@ public:
     Coordinates(int _x, int _y);
     Coordinates(const Coordinates& other);
     virtual bool operator==(const Coordinates& other) const;
-    double getCol(){ return col; }
-    double getRow(){ return row; }
+    int getCol(){ return col; }
+    int getRow(){ return row; }
 //     void setCol(double _x){ col = _x; }
 //     void setRow(double _y){ row = _y; }
